+++ specs/requests/presentations_spec.py
@@ -18,7 +18,10 @@ with description('/presentations') as self:
         current_app.es = Stub()
 
     with after.each:
-        self.context.delete()
+        # If check to prevent tests from failing occasionally
+        # Needs to be inspected!
+        if hasattr(self, "context"):
+            self.context.delete()
 
     with description('/'):
         with description('POST'):
